created products page & renamed containers folder
