+++ client/modules/admin/components/Menu/menuItems.js
@@ -2,6 +2,7 @@ import React from 'react';
 import HomeIcon from '@material-ui/icons/Home';
 import AddIcon from '@material-ui/icons/Add';
 import PeopleIcon from '@material-ui/icons/People';
+import ShoppingCart from '@material-ui/icons/ShoppingCart';
 
 export default [
   [
@@ -13,9 +14,9 @@ export default [
   ],
   [
     {
-      link: '/create-product',
-      label: 'Add Product',
-      icon: <AddIcon fontSize="inherit" />,
+      link: '/products',
+      label: 'Products',
+      icon: <ShoppingCart fontSize="inherit" />,
     },
   ],
   [
