+++ client/modules/admin/index.js
@@ -1,5 +1,6 @@
-export { default as AdminRoot } from './containers/Root';
-export { default as CreateProduct } from './containers/CreateProduct';
-export { default as Users } from './containers/Users';
+export { default as AdminRoot } from './screens/Root';
+export { default as Products } from './screens/Products';
+export { default as CreateProduct } from './screens/CreateProduct';
+export { default as Users } from './screens/Users';
 
 export { default as Sidebar } from './components/Sidebar';
