+++ client/modules/admin/screens/Products/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import Page from '../../components/Page';
+
+const Root = () => (
+  <Page title="Products">
+    <h1>Products page</h1>
+  </Page>
+);
+
+export default Root;
